Upgraded according to strict standard ~2.1.
