+++ composer.json
@@ -16,7 +16,8 @@
   },
   "require-dev": {
     "phpunit/phpunit": "~4.1",
-    "squizlabs/php_codesniffer": "~1.5"
+    "squizlabs/php_codesniffer": "~2.0",
+    "ongr/ongr-strict-standard": "~2.1"
   },
   "autoload": {
     "psr-4": {"ONGR\\MonitoringBundle\\": ""}
